Re-organized components
