+++ autoscheduler/frontend/src/App/App.css.d.ts
@@ -1,0 +1,1 @@
+export const appContainer: string;
