+++ autoscheduler/frontend/src/App/App.tsx
@@ -4,12 +4,13 @@ import { AppBar, IconButton, Toolbar } from '@material-ui/core';
 import HomeIcon from '@material-ui/icons/Home';
 import { ThemeProvider } from '@material-ui/styles';
 
-import theme from './theme';
-import Empty from './components/Empty';
-import Schedule from './components/Schedule';
-import Meeting, { MeetingType } from './types/Meeting';
-import Section from './types/Section';
-import Instructor from './types/Instructor';
+import theme from '../theme';
+import Empty from '../components/Empty/Empty';
+import Schedule from '../components/Schedule/Schedule';
+import Meeting, { MeetingType } from '../types/Meeting';
+import Section from '../types/Section';
+import Instructor from '../types/Instructor';
+import * as styles from './App.css';
 
 // DEBUG
 const testMeeting = new Meeting({
@@ -64,7 +65,7 @@ const testMeeting2 = new Meeting({
 
 const App: React.SFC = function App() {
   return (
-    <div className="app-container">
+    <div className={styles.appContainer}>
       <ThemeProvider theme={theme}>
         <AppBar position="static">
           <Toolbar>
