+++ autoscheduler/frontend/src/components/MeetingCard/MeetingCard.css
