+++ autoscheduler/frontend/src/components/MeetingCard/MeetingCard.css.d.ts
