+++ autoscheduler/frontend/src/components/Schedule/Schedule.css.d.ts
